+++ src/Common/SystemUtils.cs
@@ -15,12 +15,13 @@
 // You should have received a copy of the GNU Lesser General Public License
 // along with SCaddins.  If not, see <http://www.gnu.org/licenses/>.
 
+namespace SCaddins.Common
+{
+    
 using System;
 using System.Security;
 using System.Diagnostics;
 
-namespace SCaddins.Common
-{
     public static class SystemUtilities
     {
         [SecurityCritical]